+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+2.2.0
++++++
+* BREAKING CHANGE: Remove 'acr build-task' command group.
+* BREAKING CHANGE: Remove --tag/--manifest from 'acr repository delete' command.
+
 2.1.13
 ++++++
 * Add support for VNet/IP rules.